Move session to a separate module
