+++ test_project/pg/tests.py
@@ -4,7 +4,7 @@ from sqlalchemy import select
 from sqlalchemy.dialects.postgresql import array
 
 from aldjemy.core import get_engine
-from aldjemy.orm import get_session
+from aldjemy.session import get_session
 from test_project.pg.models import DateRangeModel, JsonModel, TicTacToeBoard
